Fix CSV downloads for IE11
